% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grasshopper_plot.R
\name{grasshopper_plot}
\alias{grasshopper_plot}
\title{grasshopper_plot}
\usage{
grasshopper_plot(df, xaxis, yaxis, group_by, fill_by, point_size = 3,
  shape = "square", do_label = FALSE, main_title = NULL,
  show_value_guide = FALSE, value_guide_title = NULL)
}
\arguments{
\item{df}{A data frame with three categorical variables and one response
value.}

\item{xaxis}{Category to plot along the x-axis.}

\item{yaxis}{Category to plot along the y-axis.}

\item{group_by}{Category to group the values by.}

\item{point_size}{Size to use for plotting values.  Default: 3.}

\item{shape}{Shape to use for the plotting values.  To see acceptable values
run the function view_shapes().  Default: 'square'}

\item{do_label}{Should each point be labeled with the grouping variable
value? Currently non-functional.  Default: FALSE}

\item{show_value_guide}{Should a colorbar legend be displayed?  Default: FALSE.}

\item{plot_title:}{Default: NULL}

\item{value_guide_title:}{Title to display for the colorbar (if displayed).
Default: "value"}
}
\description{
A function to display data that is discribed by three categorical values.
  Essentially, a grouped heatmap.  For example, one could plot the changes in the
  levels (value) of particular proteins (variable 1) in different cell types
  (variable 2) to a particular stimulus (variable 3).

  Note: since this uses ggplot2, parameters passed to xaxis, yaxis, and group_by
  should be unquoted.
}
