% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{plot_seq}
\alias{plot_seq}
\title{Plot the best and the worse}
\usage{
plot_seq(x, top = 4, bottom = 4)
}
\arguments{
\item{x}{An object of class \code{grasshopper_sim}}

\item{top}{Integer scalar}

\item{bottom}{Integer scalar}
}
\description{
Plot the best and the worse
}
