% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{grasshopper}
\alias{grasshopper}
\alias{plot.grasshopper_sim}
\title{The grasshopper problem}
\usage{
grasshopper(nsim = 10000, area = 100, len = 3, subsim = 5000,
  nchanges = 5, ring = 0.99, verb = TRUE)

\method{plot}{grasshopper_sim}(x, y = NULL, id = x$sol, ...)
}
\arguments{
\item{nsim}{Integer scalar.}

\item{area}{Integer scalar.}

\item{len}{Integer scalar.}

\item{subsim}{Integer scalar.}

\item{x}{An object of class \code{grasshopper_sim}}

\item{y}{ignored.}

\item{id}{Id of the grass to plot.}

\item{...}{Ignored.}

\item{xmax}{Integer scalar.}

\item{ymax}{Integer scalar.}
}
\value{
A list
}
\description{
The grasshopper problem
}
\examples{
1+1
}
