% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{grasshopper}
\alias{grasshopper}
\alias{plot.grasshopper_sim}
\title{The grasshopper problem}
\usage{
grasshopper(nsim = 10000, xmax = 100, ymax = 100, area = 100, len = 3,
  subsim = 5000, ring = 0.5, ncpus = parallel::detectCores())

\method{plot}{grasshopper_sim}(x, y = NULL, id = x$sol, ...)
}
\arguments{
\item{nsim}{Integer scalar.}

\item{xmax}{Integer scalar.}

\item{ymax}{Integer scalar.}

\item{area}{Integer scalar.}

\item{len}{Integer scalar.}

\item{subsim}{Integer scalar.}

\item{ring}{Numeric scalar. Size of the inner circle, if any. If set to 0
then the grasshopper jumps between its spot and \code{len} uniformly, if greater
than 1 then the grasshopper always jumps \code{len}.}

\item{ncpus}{Integer scalar.}

\item{x}{An object of class \code{grasshopper_sim}}

\item{y}{ignored.}

\item{id}{Id of the grass to plot.}

\item{...}{Ignored.}
}
\value{
A list
}
\description{
The grasshopper problem
}
\examples{
1+1
}
